Move test to different file and cleanup some
